AppList needs a context object
A sample test setup which makes the activity manager not null. The
problem is that, since it is an outside class, it can't see the context.
You need to get the context in the OnCreate() and pass it to the
constructor of AppList(). However it can't be stored in the class I
don't think. We will need to do a workaround to make the updateList()
method work. Maybe it could call the constructor again if that is
possible.